게시글/댓글 등록/수정/삭제 시 인증 정보 추가

+++ src/main/java/com/back/domain/post/post/controller/ApiV1PostController.java
@@ -52,8 +52,15 @@ public class ApiV1PostController {
     @DeleteMapping("/{id}")
     @Transactional
     @Operation(summary = "삭제")
-    public RsData<Void> deleteItem(@PathVariable long id) {
+    public RsData<Void> deleteItem(@PathVariable long id,
+                                   @NotBlank @Size(min = 2, max = 50) @RequestHeader("Authorization") String authorization) {
+        String apiKey = authorization.replace("Bearer ", "");
+        Member author = memberService.findByApiKey(apiKey)
+                .orElseThrow(() -> new ServiceException("401-1", "준재하지 않는 회원입니다."));
         Post post = postService.findById(id);
+        if (!author.equals(post.getAuthor())) {
+            throw new ServiceException("403-1", "글 삭제 권한이 없습니다.");
+        }
         postService.delete(post);
         return new RsData<>("200-1", "%d번 게시글이 삭제되었습니다.".formatted(id));
     }
@@ -82,10 +89,21 @@ public class ApiV1PostController {
     @Transactional
     @Operation(summary = "수정")
     public RsData<Void> modify(@PathVariable long id,
-                         @Valid @RequestBody PostModifyReqBody reqBody) {
+                         @Valid @RequestBody PostModifyReqBody reqBody,
+                               @NotBlank @Size(min = 2, max = 60) @RequestHeader("Authorization") String authorization) {
+
+        String apiKey = authorization.replace("Bearer ", "");
+
+        Member author = memberService.findByApiKey(apiKey).orElseThrow(
+                () -> new ServiceException("401-1", "존재하지 않는 회원입니다."));
 
         Post post = postService.findById(id);
-        postService.update(post, reqBody.subject(), reqBody.body());
+
+        if (!author.equals(post.getAuthor())) {
+            throw new ServiceException("403-1", "글 수정 권한이 없습니다.");
+        }
+
+        postService.update(post, reqBody.title(), reqBody.content());
 
         return new RsData<>(
                 "200-1",
